requirements.txt
azure.eventhub
faker
asyncio
